chat client: make host part of the config options
I think that this its a better idea to populate the host field on the server instead of guessing it on the client because guessing it on the client is quite fragile: it only works when the webmaster includes our script the way we tell them to. If that doesn't fit into their workflow (e.g. they want to concatinate all javascript files, they want to load it asynchronously) our functionality will break.