Fix printview for browsers (Issue#2162)
This patch fixes the currently broken view of ticket zoom (when you try to print a ticket). All "read more" tags will be "open" so that you get the full artcle content. The scrollbar is now hidden and you can also print all pages. 

Verified on current Chrome and Firefox. In case this brakes the print overview within the browser, I'll fix that on Monday - promise!