Fixes #2715 - Twitter status URLs are broken
This commit is a complement of its parent:
the parent fixes broken links to Twitter DMs,
while this one fixes broken links to Twitter statuses.

When importing tweets (articles of type "twitter status"),
Zammad generates a source link URL using the following format:

    https://twitter.com/statuses/:id

and then stores that URL under `article.preferences[:links]`.

This URL template worked until as recently as 2017[0],
but currently fails for users actively logged in to Twitter.
Now, the correct URL template appears to be

    https://twitter.com/:user_id/status/:id

where `:user_id` is not strict, and may be any word (\w+) <= 20 chars.
Try it yourself:

    $ curl https://twitter.com/elonmusk/status/1069382411899817990
    <html><body>You are being <a href="https://twitter.com/medenhofer/status/1069382411899817990">redirected</a>.</body></html>

In this commit, we replace `:user_id` with a single underscore (`_`).
This behavior is not officially documented anywhere (as far as I know),
but it works (for now).

This commit also extends the previous commit's DB migration/bg job
to rectify existing, broken tweet URLs stored in the database.
For performance purposes, this migration is performed in the background
and limited to the latest 10,000 Twitter articles.

[0]: https://stackoverflow.com/questions/41786123
GitHub: https://github.com/zammad/zammad/issues/2715