Maintenance: Enhance VCR helper with better failure messages
8b6c26a7b added an RSpec metadata flag to simplify the use of VCR:

    describe 'super cool feature', :use_vcr do
      it 'totally works' { ... }
    end

Under the hood, this option automatically generates filenames
for the VCR cassettes it uses on each example it's applied to.
These filenames are based on the example descriptions;
for the sample block above, the resulting filename would be

    super_cool_feature_totally_works.yml

This introduces the risk of test regressions
that could be extremely confusing and hard to debug,
simply because someone changed the example description.

This commit injects custom error messages into RSpec
to elucidate this problem and avoid needless debugging.

=== Design challenges

This error message injection uses Ruby's Module#prepend
to monkey-patch methods defined in RSpec,
meaning that the changes are coupled to RSpec's implementation.
In short, if the implementation changes,
this custom error messaging could break.

Specifically, there are two different modes of failure in RSpec,
and the custom error was thus injected in two corresponding places:

* `.notify_failure` for normal test failure
  (i.e., when an expectation is not met); and
* `.handle_matcher` for exceptions raised during a test.