Follow up - c94000d8eba92e0b77d45d6758e82add1b13c7fd - ActiveJobLock lock update fails due to PG::TRSerializationFailure, 'Could not serialize access due to concurrent update' exception
It indicates that the UPDATE or DELETE statement was queued behind another UPDATE/DELETE statement on the same row. That other statement finished, and due to the guarantees of the Serializable isolation level, the statement in this session was canceled.

The PostgreSQL states that the transaction should automatically retry the UPDATE or DELETE when seeing this error message, whilst taking the updated state of the row into account.

See 'Serializable Isolation Level' section of https://www.postgresql.org/docs/10/transaction-iso.html