Maintenance: explicitly expect zammad user and exit if not found
As the restoration script expects specific users and all package installation and source code documentation is talking about "zammad" as a user, we expect this user to be present.

If needed users that want to use different user names may have to adjust the scripts. This still should cover over 90% of Zammad installations.

This is a preparation commit for Issue #3139.