Fix article top border
When we render the security & encryption info above an article, we hide the top border and disable the border rounding at the top, so that it fits better. This layout gets enabled when the element .article-meta-permanent. So far this element was rendered, even when there was no info box to show. This broke the design of the article. This fix only adds the element when its visible.